+ changed preselected toons based on tab
please check if that suits you - for me it is more natural: if I select another toon in the dropdown at top, then go to all toons tab, I want to trade to this toon usually. If I am at bags section on a specific toon, I want to trade with me, usually.